Add more context aware option generation
If a different user is set via a -u option, then completion should be aware of that option and
cater the suggestions with that option in mind, like completing only changelists from that user
when the -e (filter changelist) option is used.

Contextual options are currently client, user, stream, and change status.  i.e. if an option
can set a designated client, user, stream, or status, the rest completion should become aware
of that.